fixed the bug reading S3 files
