+++ pom.xml
@@ -15,7 +15,7 @@
     <description>The OneDataShare transfer service.</description>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
         <spring-cloud.version>Hoxton.SR6</spring-cloud.version>
     </properties>
