+++ src/main/java/org/onedatashare/transferservice/odstransferservice/service/JobControl.java
@@ -159,6 +159,7 @@ public class JobControl extends DefaultBatchConfigurer {
     public Job concurrentJobDefinition() throws MalformedURLException {
         List<Flow> flows = createConcurrentFlow(request.getSource().getInfoList(), request.getSource().getParentInfo().getPath(), request.getJobId());
         Flow[] fl = new Flow[flows.size()];
+        threadPoolConfig.setSTEP_POOL_SIZE(this.request.getOptions().getConcurrencyThreadCount());
         Flow f = new FlowBuilder<SimpleFlow>("splitFlow").split(threadPoolConfig.stepTaskExecutor()).add(flows.toArray(fl))
                 .build();
         return jobBuilderFactory.get(request.getOwnerId()).listener(new JobCompletionListener())
